refactor: move axis-lock-projector to shared
Lets the renderer consume the same projector during option-drag copy
without crossing the main/renderer layer boundary.
