+++ src/shared/axis-lock-projector.ts
